+++ httprunner/schema/__init__.py
@@ -1,1 +1,3 @@
+from .api import Api
 from .testcase import ProjectMeta, TestCase, TestCases
+from .testsuite import TestSuite