+++ httprunner/schema/api.py
@@ -1,4 +1,4 @@
-from pydantic import BaseModel
+from pydantic import BaseModel, Field
 
 from httprunner.schema import common
 
@@ -7,8 +7,8 @@ class Api(BaseModel):
     name: common.Name
     request: common.Request
     variables: common.Variables
-    base_url: common.BaseUrl
-    setup_hooks: common.Hook
-    teardown_hooks: common.Hook
-    extract: common.Extract
-    validate: common.Validate
+    base_url: common.BaseUrl = ""
+    setup_hooks: common.Hook = []
+    teardown_hooks: common.Hook = []
+    extract: common.Extract = {}
+    validation: common.Validate = Field([], alias="validate")