+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 **Changed**
 
+- replace jsonschema validation with pydantic
 - remove compatibility with testcase/testsuite format v1
 
 ## 3.0.1 (2020-03-24)